requirements.txt
pygame~=2.1.2
peewee~=3.14.8

PyYAML~=6.0
